@@ -1,6 +1,6 @@
 {
   "name": "texta",
-  "version": "1.4.10",
+  "version": "1.4.11",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -20,25 +20,25 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/build-angular": "^0.1000.3",
-    "@angular/animations": "~10.0.4",
+    "@angular/animations": "~10.1.1",
     "@angular/cdk": "^10.0.2",
-    "@angular/common": "~10.0.4",
-    "@angular/compiler": "~10.0.4",
-    "@angular/core": "~10.0.4",
+    "@angular/common": "~10.1.1",
+    "@angular/compiler": "~10.1.1",
+    "@angular/core": "~10.1.1",
     "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "~10.0.4",
+    "@angular/forms": "~10.1.1",
     "@angular/material": "^10.0.2",
-    "@angular/platform-browser": "~10.0.4",
-    "@angular/platform-browser-dynamic": "~10.0.4",
-    "@angular/router": "~10.0.4",
-    "@swimlane/ngx-charts": "^14.0.0",
+    "@angular/platform-browser": "~10.1.1",
+    "@angular/platform-browser-dynamic": "~10.1.1",
+    "@angular/router": "~10.1.1",
+    "@swimlane/ngx-charts": "^16.0.0",
     "@types/linkify-it": "^2.1.0",
-    "angular-plotly.js": "^2.0.0",
+    "angular-plotly.js": "^3.0.0",
     "core-js": "^2.6.11",
     "http-server": "^0.12.3",
     "linkify-it": "^2.2.0",
     "ngx-infinite-scroll": "^8.0.2",
-    "ngx-mat-select-search": "^3.0.1",
+    "ngx-mat-select-search": "^3.1.0",
     "ngx-material-file-input": "^2.1.1",
     "npm-run-all": "^4.1.5",
     "plotly.js": "^1.54.5",
@@ -48,8 +48,8 @@
   },
   "devDependencies": {
     "@angular/cli": "~10.0.3",
-    "@angular/compiler-cli": "~10.0.4",
-    "@angular/language-service": "~10.0.4",
+    "@angular/compiler-cli": "~10.1.1",
+    "@angular/language-service": "~10.1.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "^12.12.38",
