@@ -54,11 +54,11 @@ import {FromToInputComponent} from './components/from-to-input/from-to-input.com
 import {ScrollTopDirective} from './directives/scroll-top.directive';
 import {PortalModule} from '@angular/cdk/portal';
 import {OverlayModule} from '@angular/cdk/overlay';
-import { PlotlyViaCDNModule } from 'angular-plotly.js';
+import {PlotlyViaCDNModule} from 'angular-plotly.js';
 
 
-PlotlyViaCDNModule.plotlyVersion = '1.49.4'; // can be `latest` or any version number (i.e.: '1.40.0')
-PlotlyViaCDNModule.plotlyBundle = 'gl2d'; // 'basic', 'cartesian', 'geo', 'gl3d', 'gl2d', 'mapbox' or 'finance'
+PlotlyViaCDNModule.setPlotlyVersion('1.49.4'); // can be `latest` or any version number (i.e.: '1.40.0')
+PlotlyViaCDNModule.setPlotlyBundle('gl2d'); // 'basic', 'cartesian', 'geo', 'gl3d', 'gl2d', 'mapbox' or 'finance'
 
 @NgModule({
   declarations: [LoginDialogComponent, GenericTableComponent,
